// types/currency.ts
export enum CurrencyCode {
    USD = 'USD',
    EUR = 'EUR',
    GBP = 'GBP',
    CAD = 'CAD',
    AUD = 'AUD',
    USDC = 'USDC',
    USDT = 'USDT',
    EURC = 'EURC'
}

export enum CurrencyFullName {
    USD = 'US Dollar',
    EUR = 'Euro',
    GBP = 'British Pound',
    CAD = 'Canadian Dollar',
    AUD = 'Australian Dollar',
    USDC = 'Digital Dollar',
    USDT = 'Tether USD',
    EURC = 'Euro Stablecoin'
}